doom emacs: remove obsolete comment about "literate" module
We already use literate programming on our own for our Doom Emacs
configuration. And since 101ec542 (emacs: enable automatic tangling on
save, 2024-08-03), we automatically tangle on save. So the "literate"
module does not add any value for us.

The comment implied that we wanted to enable it again (which we don't),
so remove it.