emacs org-mode: remove conflicting navigational bindings
We've started using magit to craft commit messages. When creating a
commit from magit, magit shows two buffers --- one with COMMIT_EDITMSG
and another below it with magit-diff. The magit-diff buffer is like a
regular text diff, but better because it supports navigating around
hunks as well as folding/unfolding hunks.

Oftentimes we want to move to the magit-diff buffer, and so we want to
use C-j to get there (as is typically the case in other major modes with
our existing setup). However this is bound to #'org-down-element by
default (we found this by typing ~,Hk~ then ~C-j~ from inside the
COMMIT_EDITMSG buffer while inside insert mode).

Remove ~C-j~ as well as other similar bindings that conflict with our
existing window/tab navigation bindings.