emacs: fix binding for commenting lines/regions
In c45bec82 (emacs: simpler binding for commenting lines/regions,
2024-08-03) we introduced the ~,c,~ binding to comment lines and
regions quickly. Unfortunately this binding did not work because it
waited until the =evil-nerd-commenter= package was loaded. And it
appears that this package is not loaded immediately, so the binding was
never available by default.

Assign the binding after the =evil= package. This makes the binding load
a bit more eagerly because we know that the =evil= package is loaded
eagerly itself.