emacs: fix shfmt invocation
We get an error like

    eval: Invalid function: "--indent"

because the setting here is not being interpreted the same as before.
This might be because the macro for "set-formatter!" changed since we
introduced this code in f4d4343c (emacs: customize formatter for shell
scripts, 2022-08-12). There appears to be a related upstream issue
saying that the docs for this macro are wrong [1], so it's plausible
that things did change in a breaking way.

We only noticed this change recently when we started installing "shfmt"
on Mac with 6e60df6d (nix-darwin: add shellcheck and shfmt, 2024-08-14).

[1] https://github.com/doomemacs/doomemacs/issues/7762