add script to detect long lines in Org files
Use babashka [1] for scripting this solution. The cool thing is that we
can use cider-jack-in-clj from Emacs to run parts of the script in the
REPL. This is much nicer than writing raw Bash.

We could have written the same script with Emacs Lisp, like this

    #!/usr/bin/env emacs -Q --script

    (defun randy ()
      (random t)
      (if (> (length command-line-args-left) 0)
          (princ (random (string-to-number (nth 0 command-line-args-left))))
        (princ (random)))
      (princ "\n"))

    (randy)

as demonstrated in [2], but it felt much nicer to use Babashka because
of its rich ecosystem specifically designed for scripting one-off
things.

[1] https://babashka.org/
[2] https://www.reddit.com/r/emacs/comments/110byrl/comment/j8941g6/