add tests for find_long_lines
Also create a test runner to run all tests. The test runner doesn't do
any sort of discovery of test files and needs to be told which tests to
run. However in the future we could fix this up so that it requires less
ceremony.

This test runner is stolen from the Babashka book [1].

[1] https://github.com/babashka/book/blame/eea70f8cf5185eb65e87066ddaf36c66fe22bceb/src/recipes.adoc#L22-L36