emacs: make org-mode pretty
This brings in org-appear [1] and org-superstar [2], on top of
org-fancy-priorities [3]. Bringing in [3] is important because it
appears that when we removed it in 6a42a636 (emacs: remove redundant
package, 2024-08-11), it was in error (we thought we had enabled the
"+pretty" module argument in Doom Emacs, but we had not).

We don't care too much about org-superstar, but org-appear is a big win
in terms of ergonomics.

[1] https://github.com/awth13/org-appear
[2] https://github.com/integral-dw/org-superstar-mode
[3] https://github.com/harrybournis/org-fancy-priorities