emacs: prettify org mode headlines and lists
This makes the headlines and lists "pop out" a bit more, especially in
terminal Emacs where things can be a little bit subdued.

Note that the customization for "org-superstar-item-bullet-alist" is an
alist, not just a plain list (which is the case for
"org-superstar-headline-bullets-list"). If we treat it like a plain
list, org-superstar-mode silently does not apply any changes to plain
lists.