tmux: remove dead code (tmux plugins)
Remove tmux plugins, because in practice we never used them.

For tmux-resurrect, our tmux sessions were simple enough that we never
really needed to reconstruct them perfectly (and actually, if we want
something that's more accurate than tmux-resurrect, we should just write
a startup script that sets things up correctly with real commands).

For extrakto, we never used it because navigating around copy-mode in
tmux is already so quick and easy.

Deleting extrakto frees up the C-M-y keybinding, so use that instead of
C-M-S-y to enter copy-mode.