emacs: insert the next "thing" based on context
This way, we can stop manually pressing RET and space/tab multiple times
to do indentation for plain list items like

- foo
  + bar
- baz

and what's more, we can also use our usual M-S-{h,l} bindings to change
the list item's indentation. As an added bonus, it appears that
indenting one level changes the list type from "-" to "+", so we can
pick up the visual cues we established in cc25835b (emacs: prettify org
mode headlines and lists, 2024-08-17) where "+" is displayed as "↪".
Although cc25835b didn't mention it, the reason why we chose to display
"↪" is because

1. we often nest one level to write down some "counterpoint" to the item
   in the first level, and

2. we want to discourage indentation beyond 2 levels for plain lists.

Lastly we add a passthrough binding in tmux because otherwise
passthrough mode swallows the C-RET binding (and we can't use C-RET in a
remote emacs session inside a nested tmux).