emacs: set tab-width to 8 for makefiles and shell files
We do this because TAB characters are traditionally 8 spaces long. If we
want tighter indentation, we should be using SPACEs, not TABs.

Also, both shell and Makefiles are not normally meant to involve complex
logic. So if we find ourselves needing to indent heavily, it's a good
sign that we need to move to a higher-level programming language.

One nice side-effect of this is that TAB characters in shell and
Makefile source code blocks in Org mode are displayed with a width of 8
spaces. This is because "org-src-font-lock-fontify-block" uses the
major mode of the language designated for the source code block to
figure out how wide TAB characters should be when displayed in the Org
buffer.

The above behavior of source code blocks inside Org mode buffers was the
original motivation for this commit.