find_long_lines (Org): remove emphasis marker truncation
The previous check of emphasis markers was there because in Org mode we
hide emphasis markers, and this actually make org-fill-paragraph treat
the line as if it does not have those markers. So we would get lines
that are just over 80 characters long. And the point of
truncate-emphasis-markers was to remove these markers to simulate what
org-fill-paragraph was seeing.

We've since increased the warning threshold to 100 columns, and so we
are rarely seeing any lines that are just over 80 columns as before due
to the emphasis markers. So remove this logic as it's unnecessary.