emacs: preserve tabs if indent-tabs-mode is set
The old invocation always passed a positive number to "--indent", which
made it always replace TABs with SPACEs. The only way to preserve tabs
is by passing in 0 for this flag.

For shell scripts where we want to use spaces, we can just disable
indent-tabs-mode (which will set the indent-tabs-mode variable to
nil).