introduce editorconfig
EditorConfig [1] can be used to maintain consistent style in a project.
We use it here to codify stylistic choices in a more official manner
than traditional editor-specific configuration files (like .dir-locals
for Emacs, for example).

Ultimately we want to move to a completely literate configuration, where
only the Org files will matter. When we do move over to a completely
literate style, this file will likely be deprecated.

[1] https://editorconfig.org/