emacs: fix notmuch sync script invocation
Although we claimed that the invocation of the sync script was fixed in
d2b522f6 (emacs: configure notmuch inside use-package! macro,
2024-08-04), that claims appears to be false. This is because when we
used the ",mu" binding to invoke (+notmuch/update), it would still use
the default (autoloaded) definition of +notmuch-get-sync-command (and
not the version that we had customized).

Use an advice wrapper so that our own definition always gets used
instead. According to [1], advice can advise functions before they are
even loaded, which is what we want.

[1] https://stackoverflow.com/a/15725437/437583