git: wrap diff portion in COMMIT_MSG with source code block markers
Also set Git's core.hooksPath to ~/syscfg/script/git/hooks to enable
this new hook (the Clojure version, which is the one symlinked as just
prepare-commit-msg) globally. We use a symlink here because otherwise
we'd have to rename our Clojure file with dashes, but using dashes is
against Clojure convention.