zsh: run fzf to find files after "d  "
In c8407b80 (zsh: bind M-f to complete files, 2024-08-30) we added a
shortcut to find files quickly with fzf. However, most of the time we
use it as an argument to "d", our alias to our universal file opening
function "univ_open". So we frequently end up doing "d " followed by M-f
to invoke fzf-file-and-enter-widget.

Instead, invoke fzf-file-and-enter-widget if we type in "d  " into the
terminal. This is much easier and quicker.