prepare-commit-msg: support additional patterns
Sometimes the comment section (just before the diff) will look a bit
different, so search for an alternate string prefix other than just "#".
Conceivably, we could add more patterns in the future as they are seen
in the wild (or use a regex, although we would prefer not to use one).