emacs orgmode: easily jump back to previous location
After clicking on an internal link, we can go back to where we were in
the buffer with org-mark-ring-goto. Somehow this function does not have
a hotkey, although Org by default have it set to "C-c &" (perhaps
evil-org or some other package overrides it).

Use "(" for this function. This overrides the existing
evil-org-backward-sentence function, but it's OK because that function
is almost useless.

In practice, using org-mark-ring-goto is very useful whenever we need to
jump to some intra-buffer location (such as by clicking on radio targets
or links to headings) only temporarily, and we want to go back to where
we were.