drop unnecessary "foreign" subdirectory
We used to clone repos not authored by us to a separate "foreign"
subdirectory under "prog". But we stopped doing this recently because it
doesn't really matter --- there just aren't enough repos we need to
clone at any given time to justify the split.

We could have done this cleanup years ago but we were too lazy.