git: only enable verbose commits for terminal Git interactions
In Magit, this option is redundant because Magit already displays a
separate window of all the diff hunks (which are collapsible with TAB).
So disable the global option.

But, still keep the option for cases when we invoke Git from the command
line (because in that case we are bypassing Magit).