emacs org-mode: mark "=" as punctuation
When we use "*" in normal mode to highlight all occurrences of the word
under the cursor (such as "foo"), text like

    =foo=

would not get highlighted, even though just

    foo

would. This is because the equal sign (=) would not get counted as
punctuation, but as a word character.

Make this character count as punctuation.

It's not clear why this character is counted as a word character by
default, but that's another matter.