orgmode: fix comment
The new comment is stolen from "org-agenda-skip-if" which discusses this
in its documentation. The "nottodo" means "Check if TODO keyword does
not match", and "done" is a keyword class that can mean any "done state"
condition, like "DONE" or "CANCELED", etc. This makes sense for the
weekly review, where we only want to see items that are finished.