org: prefer to use ID links instead of filenames or heading names
Now using org-store-link stores (read: copies to an area separate from
the clipboard) the link using a UUID if one isn't there already, and
doing org-insert-last-stored-link inserts this same link to the buffer.
The UUID is the :ID: value inside the :PROPERTIES: drawer.

This appears to work across files as well (org knows how to find the ID
even if it references another Org file). According to [1] this is
because Org stores a cache of IDs and which files they belong to in
.org-id-locations.

This also appears to work for linking to code blocks --- the format
is

  [[id:<UUID>::<name-of-block>][<name-of-block>]]

where <name-of-block> is the "#+name: ..." part of the block.

[1] https://emacs.stackexchange.com/questions/51755/org-mode-link-files-with-ids-and-not-filenames