nix-darwin: update for Sequoia 15.4
We have to tweak Emacs because native compilation is currently broken on
Nixpkgs for Mac.

"jujutsu" cannot be installed normally, because Nixpkgs has marked it as
an insecure package in Nix 24.11 (jujutsu version 0.23.0) and we make
nix-darwin use Nixpkgs version 24.11. For reference, the channels we are
using for this update are as follows:

    darwin https://github.com/nix-darwin/nix-darwin/archive/nix-darwin-24.11.tar.gz
    nixpkgs https://nixos.org/channels/nixpkgs-24.11-darwin
    nixpkgs-unstable https://nixos.org/channels/nixpkgs-unstable
    rust-overlay https://github.com/oxalica/rust-overlay/archive/master.tar.gz

To get around this, use the newer version of jujutsu from
nixpkgs-unstable (currently 0.28.2). The "import <nixpkgs-unstable>"
uses the local version of the "nixpkgs-unstable" channel. This is faster
than using builtins.fetchGit because that would require fetching the
entire Nixpkgs repo which is huge and takes several minutes on a fast
connection.

Lastly, the "environment.loginShell" bit is deprecated, so remove it,
because otherwise the expression will fail to evaluate.