tmux: update C-M-m binding to work on tmux 3.5a
It's been some time (some months?) since we've started noticing that the
C-M-m binding to rotate tmux layouts stopped working. It looks like in
3.5, there were some major changes to how keys are handled.

Prior to this commit (but on tmux 3.5a), running C-M-m would get us into
the "virt" tmux key table (our custom key table), but then run "ENTER",
which is the traditional binding for interpreting "C-m". This is the
broken behavior that previously worked on tmux 3.3a.

Get rid of the special use of the "virt" key table for "C-M-m", by
making Wezterm send tmux the csi-u binding for "C-M-m" directly. It
appears that tmux understands this natively, so there is no need to work
around the (old) tmux behavior that did not understand csi-u mode.

There are a number of bindings that we can now clean up (we could
probably just get rid of the "virt" key table now), but save that for
later. The diff here is just a minimum working example for C-M-m.