emacs: do not use org-id-link-consider-parent-id
We enabled this is f1941d80 (emacs: enable
org-id-link-consider-parent-id, 2025-05-04), but have since realized
that IDs are something we should be creating everywhere, because they
guarantee uniqueness and are easier to reason about in terms of links.
For example, it's much easier to resolve the link

    [[id:xxx][foo]]

than

    [[id:xxx:bar][foo]]

because we don't need to search for "bar" after dereferencing "xxx".
This starts to matter if we ever rename "bar" to something else in the
source, which would invalidate the second example above. But the first
example would still continue to work because the "xxx" in there
will point to the ID of the target itself, not a parent.