tig 'Y' hotkey: use committer date, not author date
The commit date is more accurate because (unless it's been artificially
tampered), it's the date when the commit was created (as part of "git
commit" or "git rebase"), not when the patch was originally authored.

"man git-show" has this to say about the "reference" format we used to
use:

    This format is used to refer to another commit in a commit message
    and is the same as --pretty='format:%C(auto)%h (%s, %ad)'. By
    default, the date is formatted with --date=short unless another
    --date option is explicitly specified.

Use the equivalent of "reference" format, but swap out author date "%ad"
with the commit date "%cd".