emacs l/scroll-jump: allow going up even at top of buffer
Because progn stops execution if there's an error, we had an edge case
where we couldn't scroll up using l/scroll-jump if the window was
displaying the start of the buffer. That is, evil-scroll-line-up would
error out because there was nothing to scroll (this function moves the
view up), and then the evil-scroll-line-up would never get executed.

Only execute evil-scroll-line-up if the window is _not_ at the beginning
of the buffer (position 1), to stop it from erroring out.

It remains a bit of a mystery how we never noticed this edge case until
now. Maybe Evil (or even Emacs itself?) changed behaviors since we last
touched this area?