wezterm: give bright black a unique color
Previously we made this the same as the background color, so sometime we
would encounter an application (like jj) which uses this and we would
get mysterious empty spots where the foreground (bright black) matched
the background.

Give bright black a unique color so that we never get these empty spots.
Tweak color settings in other applications to use a darker
foreground (or background, if that's how bright black was used) color
where we previously used to use the equivalent of bright black.