zsh: make jj complete revision names
Jujutsu has matured a bit. I have version 0.29.0 now, and according to
the online doc [1], there's a new way of configuring shell completions
for Zsh, such that it completes revisions, not just the subcommands and
help flags. This is a big productivity win.

It would be even better if jj was able to complete revisions by their
description [2].

[1] https://jj-vcs.github.io/jj/latest/install-and-setup/#dynamic_1
[2] https://github.com/jj-vcs/jj/issues/7331