zsh: add some jj aliases
We've historically preferred shell aliases over application-level
aliases, mainly because

  1) there are only a handful of commands that need to be used, even for
     something like Git, and

  2) typing shell aliases is faster (no need to have a space between the
     command and the subcommand(s)).

We've kept the same naming conventions as the Git aliases, so for
example "jca" means to modify the commit message of what's currently
"staged" (in JJ, the entire working copy is always "staged").

The alias "jig" is the equivalent of "tig", by calling out to a jj TUI
called lazyjj.

The "sync" alias is courtesy of Glen Choo, the one who introduced me to
jujutsu in the first place. It's a way to basically rebase all of your
local (unpushed) work on top of "main" branch.