tmux: fix black colors
In 3c92b4e1 (wezterm: give bright black a unique color, 2025-08-23) we
gave "black" a unique color. However we forgot to change the color for
tmux, which meant that when we were in copy-mode and did a selection, we
would get messed up black colors for the foreground text.

Update colors to avoid using "black" as a foreground; instead use the
original "bright black" color that we used to use prior to 3c92b4e1.