jj: escape code block markers and disable org-pretty-mode
Escaping code block markers with a leading comma helps to un-break the
(initial) syntax highlighting. And disabling org-pretty-mode helps us
see the actual (exact) characters that will get saved into the commit
message body (because Git is not aware of Orgmode syntax and will always
display the raw bytes of the commit message).