zsh: make "jl" show the maximum number of commits without a pager
Instead of hardcoding this to 30, use the highest number we can based on
the LINES environment variable. We have to use a function because
otherwise the LINES gets evaluated once (statically) on initial
configuration load, not dynamically at runtime.

We can use the new "jll" alias to show all commits with a pager, which
is rare but which can still come in handy.