emacs: tweak habit graph size
This is the largest we can go while still being able to view just over 80
columns of text. 80 columns is the preferred width of our buffers in
general.

We sacrifice a bit of the habit description in exchange for seeing more
history and also a bit more of the future (14 days instead of the
default of 7).